añadido manual

+++ carpeta_installer/installer.py
@@ -5,14 +5,14 @@ import zipfile
 import tempfile
 import shutil
 import ctypes
-import win32com.client
+from tkinter import messagebox as MessageBox
+import subprocess
 
 
 if not ctypes.windll.shell32.IsUserAnAdmin():
     # Si el usuario no es administrador, solicita permisos elevados y reinicia el programa con permisos de administrador
     ctypes.windll.shell32.ShellExecuteW(None, "runas", sys.executable, __file__, None, 1)
 else:
-    print(1)
     root = ctk.CTk()
     root.geometry("800x500")
     root.resizable(0,0)
@@ -55,7 +55,6 @@ else:
     ctk.CTkLabel(container3, text="Welcome to the installation wizard", font= fontBARRA).grid(row=0, column=0, padx=10, pady=10, sticky='sw')
     ctk.CTkLabel(container3, text="Click the next button to continue with the installation", font= fontTEXTcoll).grid(row=1, column=0, padx=10, pady=10, sticky='nw')
 
-
     container2 = ctk.CTkFrame(root)
     container2.grid(row=1, column=0, sticky='nsew', padx=20, pady=(0, 20))
     container2.grid_rowconfigure(0, weight=1)
@@ -77,51 +76,36 @@ else:
         try:
             # Intenta mover la carpeta temporal a la carpeta de destino
             shutil.move(ruta_carpeta_temporal+"\kallpa_app", ruta_program_files)
-            
-            ruta_archivo = ruta_program_files + "\kallpa_app\kallpa_app.exe"
+            # Ruta completa al archivo del script de PowerShell
+            script_file = resolver_ruta("script.ps1")
 
+            # Llamada a PowerShell para ejecutar el script
             try:
-                print("funcioné")
-                ruta_acceso_directo = os.environ["OneDrive"] + "\Escritorio\Kallpa Processor.lnk"
-                print(ruta_acceso_directo)
-                #ruta_acceso_directo = ruta_program_files + "\kallpa_app\Kallpa Processor.lnk"
-                # Obtener el identificador del objeto
-                try:
-                    shell = win32com.client.Dispatch("WScript.Shell")
-                    acceso_directo = shell.CreateShortcut(ruta_acceso_directo)
-                    acceso_directo.TargetPath = ruta_archivo
-                    acceso_directo.Save()
-                    shell = win32com.client.Dispatch("WScript.Shell")
-                    objeto = shell.Namespace(ruta_archivo)
-                except Exception as e:
-                    print(e)
-                print(objeto)
-                objeto_identificador = objeto.self.getobject("{PIDLISTABSOLUTE}").getpidl()[1]
-                print(objeto_identificador)
-                # Establecer el identificador del objeto en el acceso directo
-                acceso_directo = shell.CreateShortcut(ruta_acceso_directo)
-                acceso_directo.TargetPath = ruta_archivo
-                acceso_directo.WorkingDirectory = ruta_archivo
-                acceso_directo.Arguments = ""
-                acceso_directo.ItemIDList = objeto_identificador
-                acceso_directo.Save()
-            except:
-                print("no funcioné")
-                ruta_acceso_directo = os.environ["USERPROFILE"] + "\OneDrive\Escritorio\KallpaProcessor.lnk"
-                print(ruta_acceso_directo)
-                shell = win32com.client.Dispatch("WScript.Shell")
-                acceso_directo = shell.CreateShortCut(ruta_acceso_directo)
-                acceso_directo.Targetpath = ruta_archivo
-                acceso_directo.save()
+                result = subprocess.run(['powershell.exe', '-File', script_file], capture_output=True, text=True)
+                MessageBox.showinfo(title="Conectado", message="the installation has finished successfully")
+            except Exception as e:
+                MessageBox.showinfo(title="Conectado", message="There was a problem" + e)
+
             
         except Exception as e:
-            shutil.rmtree(ruta_program_files+"\kallpa_app")
             print(e)
+            respuesta = MessageBox.askyesno(message="Ya existe una instalación anterior, ¿desea eliminarla y sobreescribir?", title="Alerta")
+            if respuesta == True:
+                shutil.rmtree(ruta_program_files+"\kallpa_app")
+                shutil.move(ruta_carpeta_temporal+"\kallpa_app", ruta_program_files)
+                script_file = resolver_ruta("script.ps1")
+                try:
+                    result = subprocess.run(['powershell.exe', '-File', script_file], capture_output=True, text=True)
+
+                    MessageBox.showinfo(title="Conectado", message="the installation has finished successfully")
+                except Exception as e:
+                    MessageBox.showinfo(title="Conectado", message="There was a problem" + e)
+            else:
+                pass
+        
             
         print(ruta_program_files, ruta_carpeta_temporal, ruta_program_files)
     
-    
-
     root.mainloop()
